requirements.txt
Pillow>=12.0.0
icalendar>=6.0.0
pytz>=2024.1
